requirements.txt
Flask~=2.0.2
twilio~=7.3.0
python-dotenv~=0.19.1
gunicorn~=20.1.0
dnspython~=2.1.0
pymongo~=3.12.1
requests~=2.26.0